# o/devicestate,many: introduce DeviceManager.preloadGadget for EarlyConfig