this makes available the config defaults from the verified
to-be-seeded gadget to set relevant early-config options

it's based on a caching loadDeviceSeed and related changes to the
seeding code

TestPopulateFromSeedCore20RunModeUserServiceTasks would have failed without the changes, as it needs the experimental flags already set when creating the seed change itself which is before the default configuration from the gadget is processed

This keeps reporting timings for the loading of assertions together under --ensure=seed:  using one of the minimal firstboot spread tests for reference,

with this code:

```
classic-snapd-firstboot# snap debug timings --ensure=seed
ID    Status        Doing      Undoing  Summary
seed                825ms            -
 ^                   37ms            -    early import assertions from seed
 ^                  777ms            -    populate state from seed
  ^                 501ms            -      load verified snap metadata from seed
   ^                187ms            -        hash and derive side info for snap "snapd"
   ^                313ms            -        hash and derive side info for snap "core18"
```

on master:

```
classic-snapd-firstboot# snap debug timings --ensure=seed
ID    Status        Doing      Undoing  Summary
seed                831ms            -
 ^                  831ms            -    populate state from seed
  ^                  41ms            -      import assertions from seed
  ^                 191ms            -      hash and derive side info for snap "snapd"
  ^                 317ms            -      hash and derive side info for snap "core18"
```