@pedronis 2021-02-04
based on #9900