@pedronis 2021-02-05
this ATM breaks broken-seeding because the behavior in the presence of a broken seed is changed, I think I know how to make us backward compatible again